+++ src/test/java/com/example/wallet_service/controller/WalletControllerTest.java
@@ -2,6 +2,7 @@ package com.example.wallet_service.controller;
 
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -11,10 +12,13 @@ import java.util.NoSuchElementException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
+import com.example.wallet_service.dto.TransactionInfo;
 import com.example.wallet_service.service.WalletService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 @WebMvcTest(WalletController.class)
 class WalletControllerTest {
@@ -48,5 +52,24 @@ class WalletControllerTest {
             .andExpect(status().isNotFound())
             .andExpect(jsonPath("$.message").value("User not found"))
             .andExpect(jsonPath("$.status").value(404));
+    }
+    
+    @Test
+    void debit_reduceBalance() throws Exception {
+        String userId = "user123";
+        double balance = 100.0;
+        double amount = 50.0;
+        TransactionInfo transactionInfo = new TransactionInfo(userId, amount);
+
+        when(walletService.getBalance(userId)).thenReturn(balance);
+        when(walletService.debit(userId, amount)).thenReturn(amount);
+            
+        mockMvc.perform(
+                post("/transactions/debit", transactionInfo)
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(new ObjectMapper().writeValueAsString(transactionInfo))
+            )
+            .andExpect(status().isOk())
+            .andExpect(content().string("{\"balance\":50.0}"));
     }
 }